deploy/aws/us-east-2/requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
PyJWT==2.8.0
PyOpenSSL==23.2.0
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
urllib3<2
psycopg==3.1.10
psycopg-binary==3.1.10
pycparser==2.21
typing_extensions==4.7.1
